refactor: Move bookmark utils from shared-react to shared
