+++ apps/web/components/dashboard/bookmarks/BookmarkCard.tsx
@@ -1,7 +1,7 @@
 import { api } from "@/lib/trpc";
 
-import { isBookmarkStillLoading } from "@karakeep/shared-react/utils/bookmarkUtils";
 import { BookmarkTypes, ZBookmark } from "@karakeep/shared/types/bookmarks";
+import { isBookmarkStillLoading } from "@karakeep/shared/utils/bookmarkUtils";
 
 import AssetCard from "./AssetCard";
 import LinkCard from "./LinkCard";
