+++ packages/shared/utils/assetUtils.ts
